assets: Add `wp_register_script` wrapper (and then use it everywhere) (#21689)
Our Webpack-built scripts tend to need a bunch of boilerplate:
 - A call to `Assets::get_file_url_for_environment()` for possible
   debugging.
 - A call to `wp_register_style()` for extracted CSS. Possibly with
   detection of RTL.
 - Loading of dependencies and version provided by
   `@wordpress/dependency-extraction-webpack-plugin`.
 - Avoiding WPCom's broken minifier.

The wrapper handles all of that.

This also updates the loading of Webpack-built scripts to use it.
Jetpack has a bunch of different ways scripts get built. This only
touches the ones that are built with our webpack config.

This doesn't touch Jetpack's admin.js or the blocks' view.js. For the
former, the JS and CSS are being registered in different places and the
CSS is being enqueued twice for some reason. For the latter, there's a
bunch of extra logic for handling previews that I didn't want to mess
with in this PR.

Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1463419732