assets: Provide the `wp-jp-i18n-state` script (#22061)
This script is the one expected by the default configuration of the new
`@automattic/i18n-loader-webpack-plugin`.

Note the `domainMap` is currently empty. Code to fill that in will come
later, as part of #21690, once we figure out the exact form of it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1559519349